+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## 2021-08-14
 
+## Added
+
+- **Lithuanian** Added QTH locator map
+
 ## Changed
 
 - Mapbox updated to latest version